patchClient

+++ src/main/java/fr/groupe_3/fil_rouge/dao/ClientRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import fr.groupe_3.fil_rouge.entity.Client;
 
-public interface ClientRepo extends JpaRepository<Client, Long> {
+public interface ClientRepo extends JpaRepository<Client, Integer> {
 
 }
